// shared/types.ts
export type MySQLResultRows = Array<any> & { insertId: number };
export type MySQLColumnCatalogs = Array<any>;
export type MySQLResultSet = [MySQLResultRows, MySQLColumnCatalogs];

export interface MySQLQueryable {
  query(sql: string, param?: ReadonlyArray<any>): Promise<MySQLResultSet>;
}

export interface MySQLClient extends MySQLQueryable {
  beginTransaction(): Promise<void>;
  commit(): Promise<void>;
  rollback(): Promise<void>;
  release(): Promise<void>;
}

/**
 * Structure of `member` table
 */
export interface Member {
  slackID: string;
  name: string;
  kana: string;
  grade: string;
}

/**
 * Simple version of `member` table
 */
export interface MemberSimple {
  id: string;
  name: string;
}

/**
 * Structure of `easteregg` table
 */
export interface Easteregg {
  id: string;
  count: number;
  mentions: number;
}